refactor(test): add shared utilities layer and remove dead code
- Add src/tests/common/ with shared types, seed functions, and exports
- Fix layer violation: integration fixtures now use common layer
- Remove dead code: batching.ts, firebaseCalls.ts (unused utilities)
- Remove story references from test files and documentation
- Clean up test fixture comments and headers
