+++ src/router.tsx
@@ -3,7 +3,16 @@ import {createRouter} from '@tanstack/react-router'
 import type {RouterContext} from './routes/__root'
 import {routeTree} from './routeTree.gen'
 
-const queryClient = new QueryClient()
+const queryClient = new QueryClient({
+	defaultOptions: {
+		queries: {
+			staleTime: 1000 * 60 * 5, // 5 minutes
+			gcTime: 1000 * 60 * 30, // 30 minutes
+			refetchOnWindowFocus: false,
+			retry: 1
+		}
+	}
+})
 
 export function getRouter() {
 	const router = createRouter({
