+++ src/tests/e2e/integration-bridge.spec.ts
@@ -1,7 +1,7 @@
 /**
  * Integration Bridge E2E Tests - Fixture Compatibility Validation
  *
- * Story 0-7, Task 0: Validate integration fixtures work in E2E context.
+ * Validates integration fixtures work in E2E context.
  * This is a GATE test - must pass before proceeding to full migration.
  *
  * Purpose:
@@ -10,9 +10,6 @@
  * - Prove Firestore fixture can seed and clear test data
  *
  * Exit Criteria: Single integration-backed E2E test passes reliably (5/5 local, 3/3 CI)
- *
- * @see Story 0-7: E2E Test Mock Elimination
- * @see Story 0-6: Integration Test Layer (fixture source)
  */
 
 import {test as base, expect} from '../fixtures'
