+++ src/tests/e2e/passkey.spec.ts
@@ -461,12 +461,12 @@ test.describe('Passkey Network Error Handling', () => {
  */
 
 /**
- * Testing Strategy Note (Post-Story 0-7):
+ * Testing Strategy Note:
  *
- * E2E tests now follow the two-tier testing pattern:
- * ✓ Server functions hit Firebase emulators directly (no success path mocks)
- * ✓ Session injection via injectSessionCookie (acceptable per Playwright best practices)
- * ✓ Virtual authenticator via CDP for WebAuthn testing
+ * E2E tests follow the two-tier testing pattern:
+ * - Server functions hit Firebase emulators directly (no success path mocks)
+ * - Session injection via injectSessionCookie (acceptable per Playwright best practices)
+ * - Virtual authenticator via CDP for WebAuthn testing
  *
  * Retained mocks (error simulation only):
  * - mockServerFunctionError() for network failures and server errors
